requirements.txt
# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# AI / LLM
openai>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
boto3>=1.34.0

# Google Cloud
google-cloud-storage>=2.16.0
google-cloud-secret-manager>=2.20.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
html2text>=2024.2.26
beautifulsoup4>=4.12.0

# Audio / Video
moviepy>=1.0.3
pydub>=0.25.1

# Auth & Utilities
PyJWT>=2.8.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
requests>=2.31.0

# Microsoft Graph API (OneDrive integration)
msal>=1.28.0
